refacotr: fetchTokenMetadata moved to utils and changed name to fetchTokenErc721Metadata
